==> rewrite/10-summary.ipynb <==
import os
import pandas as pd

result_path = "../results/07-baseline"

exps = os.listdir(result_path)



### Only time-series data results

def get_results_baseline(result_path: str):
    result_dists = []
    exps = os.listdir(result_path)
    for exp in exps:
        result_dict = pd.read_pickle(os.path.join(result_path, exp))
        result_dict["hidden_units"], result_dict["layer"], result_dict["task"], result_dict["iteration"] = exp.split("-")[:4]
        result_dict["model_category"] = "GRU"
        result_dists.append(result_dict)
    result = pd.DataFrame(result_dists)
    result.sort_values(by=["task", "auc"], inplace=True, ascending=False)
    # top_resutls = result.drop_duplicates(subset=["task", "layer"])
    summary_results = result.groupby(by=["task", "layer"]).agg({"auc":['mean','std'], "auprc":['mean','std'], "F1":['mean','std'], 'acc':['mean','std']})
    summary_results.columns = summary_results.columns.map('_'.join).str.strip('_')
    summary_results.reset_index(inplace=True)
    summary_results["model_category"] = "GRU"
    for col in ['auc_mean', 'auc_std', 'auprc_mean', 'auprc_std',
       'F1_mean', 'F1_std', 'acc_mean', 'acc_std']:
       summary_results[col] = summary_results[col] * 100
    task_mapping = {
        "los_3": "LOS > 3 Days",
        "los_7": "LOS > 7 Days",
        "mort_hosp": "In-Hospital Mortality",
        "mort_icu": "In-ICU Mortality"
    }
    for k, v in task_mapping.items():
        summary_results.loc[summary_results["task"] == k, "task"] = v
    for matric in ["auc", "auprc", "F1", "acc"]:
        summary_results[matric] = summary_results[[f"{matric}_mean", f"{matric}_std"]].apply(lambda x: "{:.2f} +/- {:.3f}".format(x[0], x[1]), axis=1)
    summary_results = summary_results[["task", "model_category","layer", "auc", "auprc", "F1"]]
    return result, summary_results
    
resutls_07, summary_results_07 = get_results_baseline( "../results/07-baseline")
summary_results_07

### Baseline time-series and NLP data results

def get_results_multimodal(result_path: str):
    result_dists = []
    exps = os.listdir(result_path)
    for exp in exps:
        result_dict = pd.read_pickle(os.path.join(result_path, exp))
        result_dict["layer"], result_dict["hidden_units"], result_dict["embedding"], result_dict["task"], result_dict["iteration"]  = exp.split("-")[:5]
        result_dict["model_category"] = "Average Moltimodal"
        result_dists.append(result_dict)
    result = pd.DataFrame(result_dists)
    result.sort_values(by=["task", "auc"], inplace=True, ascending=False)
    summary_results = result.groupby(by=["task", "embedding"]).agg({"auc":['mean','std'], "auprc":['mean','std'], "F1":['mean','std'], 'acc':['mean','std']})
    summary_results.columns = summary_results.columns.map('_'.join).str.strip('_')
    summary_results.reset_index(inplace=True)
    summary_results["model_category"] = "Average Moltimodal"
    for col in ['auc_mean', 'auc_std', 'auprc_mean', 'auprc_std',
       'F1_mean', 'F1_std', 'acc_mean', 'acc_std']:
       summary_results[col] = summary_results[col] * 100
    task_mapping = {
        "los_3": "LOS > 3 Days",
        "los_7": "LOS > 7 Days",
        "mort_hosp": "In-Hospital Mortality",
        "mort_icu": "In-ICU Mortality"
    }
    for k, v in task_mapping.items():
        summary_results.loc[summary_results["task"] == k, "task"] = v
    for matric in ["auc", "auprc", "F1", "acc"]:
        summary_results[matric] = summary_results[[f"{matric}_mean", f"{matric}_std"]].apply(lambda x: "{:.2f} +/- {:.3f}".format(x[0], x[1]), axis=1)
    summary_results = summary_results[["task", "model_category","embedding", "auc", "auprc", "F1"]]
    return result, summary_results

resutls_08, summary_results_08 = get_results_multimodal( "../results/08-multimodal")
summary_results_08

## Proposed models

def get_results_proposed(result_path: str):
    result_dists = []
    exps = os.listdir(result_path)
    for exp in exps:
        result_dict = pd.read_pickle(os.path.join(result_path, exp))
        result_dict["sequence_name"], result_dict["hidden_unit_size"], result_dict["embedding"], result_dict["task"], result_dict["iteration"], result_dict["type_of_ner"] = exp.split("-")[:6]
        result_dict["model_category"] = "Proposed Model"
        result_dists.append(result_dict)
    result = pd.DataFrame(result_dists)
    result.sort_values(by=["task", "embedding", "iteration"], inplace=True, ascending=False)
    summary_results = result.groupby(by=["task", "embedding"]).agg({"auc":['mean','std'], "auprc":['mean','std'], "F1":['mean','std'], 'acc':['mean','std']})
    summary_results.columns = summary_results.columns.map('_'.join).str.strip('_')
    summary_results.reset_index(inplace=True)
    summary_results["model_category"] = "Proposed Model"
    for col in ['auc_mean', 'auc_std', 'auprc_mean', 'auprc_std',
       'F1_mean', 'F1_std', 'acc_mean', 'acc_std']:
       summary_results[col] = summary_results[col] * 100
    task_mapping = {
        "los_3": "LOS > 3 Days",
        "los_7": "LOS > 7 Days",
        "mort_hosp": "In-Hospital Mortality",
        "mort_icu": "In-ICU Mortality"
    }
    for k, v in task_mapping.items():
        summary_results.loc[summary_results["task"] == k, "task"] = v
    for matric in ["auc", "auprc", "F1", "acc"]:
        summary_results[matric] = summary_results[[f"{matric}_mean", f"{matric}_std"]].apply(lambda x: "{:.2f} +/- {:.3f}".format(x[0], x[1]), axis=1)
    summary_results = summary_results[["task", "model_category","embedding", "auc", "auprc", "F1"]]
    return result, summary_results

result_09, summary_results_09 = get_results_proposed("../results/09-cnn")
summary_results_09

## Best baseline models

summary_results = pd.concat([summary_results_07, summary_results_08], axis=0)


# table 3
summary_results.sort_values(by=["task", "model_category", "embedding"], inplace=True)
summary_results = summary_results[["task", "model_category", "layer", "embedding", 'auc', 'auprc', 'F1']]
summary_results.to_csv("../for_report/table_3.csv", index=False)

# table 4
summary_results_2 = summary_results.groupby(by=["task"]).agg({"auc":"max", "auprc": "max", "F1": "max"}).reset_index()
summary_results_2["model_category"] = "Best Baseline"
summary_results_2["embedding"] = ""
summary_results_2 = pd.concat([summary_results_2, summary_results_09], axis=0)
summary_results_2.sort_values(by=["task", "model_category", "embedding"], ascending=False, inplace=True)
summary_results_2 = summary_results_2[["task", "model_category", "embedding", 'auc', 'auprc', 'F1']]
summary_results_2.to_csv("../for_report/table_4.csv", index=False)

resutls = pd.concat([resutls_07, resutls_08], axis=0)
resutls.sort_values(by=["task", "model_category", "auc"], inplace=True, ascending=False)
top_resutls = resutls.drop_duplicates(subset=["task", "model_category"])
print(top_resutls.to_string(index=False))
